requirements.txt
requests
psycopg2-binary  
pyTelegramBotAPI 
unittest2
schedule
pandas
  